+++ browser/data-browser/src/components/Button.tsx
@@ -195,7 +195,7 @@ export const ButtonSubtle = styled(ButtonDefault)`
   --button-text-color: ${p => p.theme.colors.textLight};
   --button-text-color-hover: ${p => p.theme.colors.main};
 
-  box-shadow: ${p => p.theme.boxShadow};
+  box-shadow: ${p => (p.theme.darkMode ? 'none' : p.theme.boxShadow)};
 `;
 
 export const ButtonAlert = styled(ButtonDefault)`