+++ browser/data-browser/src/components/ResourceUsage/ResourceUsage.tsx
@@ -1,20 +1,21 @@
-import { Resource, classes, useCollection } from '@tomic/react';
+import { Resource, core } from '@tomic/react';
 
 import { PropertyUsage } from './PropertyUsage';
-import { UsageCard } from './UsageCard';
 import { ClassUsage } from './ClassUsage';
 import { ChildrenUsage } from './ChildrenUsage';
+import { Column } from '../Row';
+import { ReferenceUsage } from './ReferenceUsage';
 
 interface ResourceUsageProps {
   resource: Resource;
 }
 
 export function ResourceUsage({ resource }: ResourceUsageProps): JSX.Element {
-  if (resource.hasClasses(classes.property)) {
+  if (resource.hasClasses(core.classes.property)) {
     return <PropertyUsage resource={resource} />;
   }
 
-  if (resource.hasClasses(classes.class)) {
+  if (resource.hasClasses(core.classes.class)) {
     return <ClassUsage resource={resource} />;
   }
 
@@ -22,22 +23,10 @@ export function ResourceUsage({ resource }: ResourceUsageProps): JSX.Element {
 }
 
 function BasicUsage({ resource }: ResourceUsageProps): JSX.Element {
-  const { collection } = useCollection({
-    value: resource.getSubject(),
-  });
-
   return (
-    <>
+    <Column>
       <ChildrenUsage resource={resource} />
-      <UsageCard
-        collection={collection}
-        title={
-          <span>
-            <strong>{collection.totalMembers}</strong> resources reference{' '}
-            {resource.title}
-          </span>
-        }
-      />
-    </>
+      <ReferenceUsage resource={resource} />
+    </Column>
   );
 }