+++ browser/data-browser/src/components/ResourceUsage/UsageRow.tsx
@@ -1,23 +1,17 @@
-import {
-  Collection,
-  classes,
-  unknownSubject,
-  useMemberFromCollection,
-} from '@tomic/react';
+import { commits, unknownSubject, useResource } from '@tomic/react';
 
 import { ResourceInline } from '../../views/ResourceInline';
 import { styled } from 'styled-components';
 import { ErrorLook } from '../ErrorLook';
 
 interface UsageRowProps {
-  collection: Collection;
-  index: number;
+  subject: string;
 }
 
-export function UsageRow({ collection, index }: UsageRowProps): JSX.Element {
-  const resource = useMemberFromCollection(collection, index);
+export function UsageRow({ subject }: UsageRowProps): JSX.Element {
+  const resource = useResource(subject);
 
-  if (resource.getSubject() === unknownSubject) {
+  if (subject === unknownSubject) {
     return (
       <ListItem>
         <ErrorLook>Insufficient rights to view resource</ErrorLook>
@@ -25,24 +19,25 @@ export function UsageRow({ collection, index }: UsageRowProps): JSX.Element {
     );
   }
 
-  if (resource.hasClasses(classes.commit)) {
+  if (resource.hasClasses(commits.classes.commit)) {
     return <></>;
   }
 
   return (
     <ListItem>
-      <ResourceInline subject={resource.getSubject()} />
+      <ResourceInline subject={subject} />
     </ListItem>
   );
 }
 
 const ListItem = styled.li`
+  display: flex;
+  align-items: center;
   list-style: none;
   padding: 0.5rem 1rem;
   border-radius: ${({ theme }) => theme.radius};
-
-  margin-left: 0;
-
+  margin: 0;
+  height: 3rem;
   &:nth-child(odd) {
     background-color: ${({ theme }) => theme.colors.bg1};
   }