+++ browser/data-browser/src/helpers/containers.ts
@@ -1,2 +1,3 @@
 export const MAIN_CONTAINER = 'main';
 export const ALL_PROPS_CONTAINER = 'all-props';
+export const CARD_CONTAINER = 'card';